requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.2
google-cloud-speech>=2.0.0
pytest
google-generativeai
python-dotenv
click
